refactored the building of the URL for flexibility.
The URL to the fhir server is now configured based on the following env
variables:

* FHIR_PROTO (default: "http"): protocol to use
* FHIR_ADDR OR HAPI_PORT_8080_TCP_ADDR (default: "localhost"): the
address to the server.  HAPI_PORT_8080_TCP_ADDR will take precedence
over FHIR_ADDR if present.
* FHIR_PORT (default: "8080"): port to use
* FHIR_PATH (default: "/fhir/baseDstu2/"): the path to use

making the final path:

FHIR_PROTO+'://' + FHIR_ADDR + ':'+FHIR_PORT+FHIR_PATH


If the protocol is http and port is 80, port will not be included in teh
final URL.  Similarly if protocol is https and port is 443 the  port
will not be included in the final URL.